Make a tasktypes directory and move all vendoring logic into a vendoring tasktype
